Try again with static casts on windows

+++ arangod/Aql/AqlItemMatrix.h
@@ -102,7 +102,7 @@ class AqlItemMatrix {
 
   size_t _nrRegs;
 
-  uint32_t _lastShadowRow;
+  size_t _lastShadowRow;
 };
 
 }  // namespace aql
